pyproject.toml
[project]
name = "filestag"
version = "0.1.0"
description = "A fast and efficient file access library for local and cloud storage"
authors = [
    {name = "Michael Ikemann", email = "michael.ikemann@gmail.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<4"
dependencies = [
    "pydantic (>=2.0.0)",
    "requests (>=2.28.0)",
    "aiofiles (>=24.0.0)",
    "httpx (>=0.27.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]

[project.optional-dependencies]
azure = ["azure-storage-blob (>=12.0.0)"]
dev = ["pytest (>=7.4.0)", "pytest-cov (>=4.0.0)", "pytest-asyncio (>=1.0.0)", "mypy (>=1.0.0)", "ruff (>=0.1.0)"]
all = ["azure-storage-blob (>=12.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^4.0.0"
pytest-asyncio = "^1.0.0"
mypy = "^1.0.0"
ruff = "^0.1.0"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.mypy]
python_version = "3.12"
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["filestag"]
omit = ["filestag/azure/*"]

[tool.coverage.report]
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
